requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
tensorflow==2.18.0
numpy==1.26.4
opencv-python-headless==4.11.0.86
pillow==10.3.0
nest-asyncio==1.5.6
python-multipart==0.0.9
